+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- *Breaking* OutputMgr: expose wl_output global id
+
 ## 0.1.0 -- 2018-04-26
 
 Initial version, including: